moved calendar widget
