+++ src/components/sidebar-right/widgets/CalendarWidget.jsx
@@ -1,15 +1,63 @@
-import React from "react";
+import * as React from "react";
+import dayjs from "dayjs";
 
-import { CardContent, Typography } from "@mui/material";
+import { Box, TextField } from "@mui/material";
+import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
+import { LocalizationProvider } from "@mui/x-date-pickers/LocalizationProvider";
+import { StaticDatePicker } from "@mui/x-date-pickers/StaticDatePicker";
 
-function CalendarWidget() {
+export default function ActionBarComponentProps() {
+	const [value, setValue] = React.useState(() => dayjs());
+	const DATE_SIZE = 25;
+
 	return (
-		<CardContent>
-			<Typography gutterBottom variant="h6">
-				Calendar widget
-			</Typography>
-		</CardContent>
+		<LocalizationProvider dateAdapter={AdapterDayjs}>
+			<Box
+				sx={{
+					"& > div": {
+						minWidth: DATE_SIZE * 9,
+					},
+					"& > div > div, & > div > div > div, & .MuiCalendarPicker-root": {
+						width: DATE_SIZE * 6.6,
+					},
+					"& .MuiTypography-caption": {
+						width: DATE_SIZE,
+						margin: 0,
+					},
+					"& .PrivatePickersSlideTransition-root": {
+						minHeight: DATE_SIZE * 7,
+					},
+					'& .PrivatePickersSlideTransition-root [role="row"]': {
+						margin: 0,
+					},
+					"& .MuiPickersDay-dayWithMargin": {
+						margin: 0,
+					},
+					"& .MuiPickersDay-root": {
+						width: DATE_SIZE,
+						height: DATE_SIZE,
+					},
+					"& .MuiPickerStaticWrapper-root": {
+						minWidth: "auto",
+					},
+					"& .MuiCalendarOrClockPicker-root  > div": {
+						width: "auto",
+					},
+				}}
+			>
+				<StaticDatePicker
+					displayStaticWrapperAs="desktop"
+					onChange={(newValue) => setValue(newValue)}
+					renderInput={(params) => <TextField {...params} />}
+					showToolbar={false}
+					value={value}
+					componentsProps={{
+						actionBar: {
+							actions: ["disable"],
+						},
+					}}
+				/>
+			</Box>
+		</LocalizationProvider>
 	);
 }
-
-export default CalendarWidget;
